requirements.txt
fastapi==0.109.2
uvicorn==0.23.2
sqlalchemy==2.0.23
pymysql==1.1.0
pydantic==2.5.2
scikit-learn==1.3.2
pandas==2.1.1
streamlit==1.26.0
requests==2.32.1
python-dotenv==1.0.0
